Connect created sites to site

+++ controllers/site.js
@@ -1,5 +1,6 @@
 const siteModel = require('../models/site')
 const jwt = require('jsonwebtoken');
+const access = require('./access')
 
 function getUniqueName(baseName) {
     return baseName.trim().toLowerCase().replace(/\s+/g, '-');
@@ -17,6 +18,11 @@ module.exports.createSite = async(req, res) => {
     }
 
     const decodedToken = getToken(req.body.token);
+    
+    allowed = access.userOnly(decodedToken.user, req.params.id)
+    if (!allowed)
+        return res.status(500).send("Not authorized");
+
     const site = new siteModel({
         name: req.body.name,
         uniqueName: uniqueName,
@@ -28,6 +34,11 @@ module.exports.createSite = async(req, res) => {
 }
 
 module.exports.deleteSite = (req, res) => {
+    const decodedToken = getToken(req.body.token);
+    allowed = access.userOnly(decodedToken.user, req.params.userID)
+    if (!allowed)
+        return res.status(500).send("Not authorized");
+
     siteModel.findByIdAndDelete(req.params.id).then(docs => {
         if (docs == null)
             return res.status(404).send("Site not found")
@@ -36,8 +47,17 @@ module.exports.deleteSite = (req, res) => {
     })
 }
 
-// TODO: ao atualizar o nome, deve-se atualizar o uniqueName
 module.exports.editSiteName = (req, res) => {
+    if ('name' in req.body) {
+        let uniqueName = getUniqueName(req.body.name)
+        req.body['uniqueName'] = uniqueName
+    }
+
+    const decodedToken = getToken(req.body.token);
+    allowed = access.userOnly(decodedToken.user, req.params.userID)
+    if (!allowed)
+        return res.status(500).send("Not authorized");
+
     siteModel.findByIdAndUpdate(req.params.id, req.body, (err, docs) => {
         if (docs == null)
             return res.status(404).send("Site not found")
